code/app/requirements.txt
streamlit
streamlit_metrics
streamlit-webrtc
pandas
numpy
seaborn
matplotlib
scikit-learn
opencensus-ext-azure
azure-keyvault
azure-identity
azure-ai-textanalytics
azure-cognitiveservices-speech
azure-cognitiveservices-vision-computervision
opencv-python
librosa
SoundFile
